writing-practice/cdk/app/requirements.txt
Pillow==11.0.0
streamlit==1.40.2
python-certifi-win32==1.6.1
openai>=1.12.0
python-dotenv==1.0.1
httpx>=0.24.1